DB changes, stock

+++ src/main/resources/carport.sql
@@ -25,9 +25,9 @@ DROP TABLE IF EXISTS `city`;
 /*!40101 SET @saved_cs_client     = @@character_set_client */;
 /*!50503 SET character_set_client = utf8mb4 */;
 CREATE TABLE `city` (
-  `city` varchar(45) NOT NULL,
-  `postal_code` varchar(45) NOT NULL,
-  PRIMARY KEY (`city`)
+                        `city` varchar(45) NOT NULL,
+                        `postal_code` varchar(45) NOT NULL,
+                        PRIMARY KEY (`city`)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb3;
 /*!40101 SET character_set_client = @saved_cs_client */;
 
@@ -39,16 +39,16 @@ DROP TABLE IF EXISTS `parts_list`;
 /*!40101 SET @saved_cs_client     = @@character_set_client */;
 /*!50503 SET character_set_client = utf8mb4 */;
 CREATE TABLE `parts_list` (
-  `part_list_id` int NOT NULL AUTO_INCREMENT,
-  `request_id` int NOT NULL,
-  `stock_id` int NOT NULL,
-  `line_price` int NOT NULL,
-  `amount` int NOT NULL,
-  PRIMARY KEY (`part_list_id`),
-  KEY `fk_parts_list_request1_idx` (`request_id`),
-  KEY `fk_parts_list_stock1_idx` (`stock_id`),
-  CONSTRAINT `fk_parts_list_request1` FOREIGN KEY (`request_id`) REFERENCES `request` (`request_id`),
-  CONSTRAINT `fk_parts_list_stock1` FOREIGN KEY (`stock_id`) REFERENCES `stock` (`stock_id`)
+                              `part_list_id` int NOT NULL AUTO_INCREMENT,
+                              `request_id` int NOT NULL,
+                              `stock_id` double NOT NULL,
+                              `line_price` double NOT NULL,
+                              `amount` int NOT NULL,
+                              PRIMARY KEY (`part_list_id`),
+                              KEY `fk_parts_list_request1_idx` (`request_id`),
+                              KEY `fk_parts_list_stock1_idx` (`stock_id`),
+                              CONSTRAINT `fk_parts_list_request1` FOREIGN KEY (`request_id`) REFERENCES `request` (`request_id`),
+                              CONSTRAINT `fk_parts_list_stock1` FOREIGN KEY (`stock_id`) REFERENCES `stock` (`stock_id`)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb3;
 /*!40101 SET character_set_client = @saved_cs_client */;
 
@@ -60,24 +60,24 @@ DROP TABLE IF EXISTS `request`;
 /*!40101 SET @saved_cs_client     = @@character_set_client */;
 /*!50503 SET character_set_client = utf8mb4 */;
 CREATE TABLE `request` (
-  `request_id` int NOT NULL AUTO_INCREMENT,
-  `length_cp` int NOT NULL,
-  `width_cp` int NOT NULL,
-  `length_rr` int NOT NULL,
-  `width_rr` int NOT NULL,
-  `roof_mat` varchar(255) NOT NULL,
-  `wood_cladding_mat` varchar(255) NOT NULL,
-  `customer_id` int NOT NULL,
-  `admin_id` int DEFAULT NULL,
-  PRIMARY KEY (`request_id`),
-  KEY `fk_design_roof1_idx` (`roof_mat`),
-  KEY `fk_design_bekledning1_idx` (`wood_cladding_mat`),
-  KEY `fk_request_user2_idx` (`admin_id`),
-  KEY `fk_request_user1_idx` (`customer_id`),
-  CONSTRAINT `fk_design_roof1` FOREIGN KEY (`roof_mat`) REFERENCES `roof` (`roof_mat`),
-  CONSTRAINT `fk_design_wood_cladding1` FOREIGN KEY (`wood_cladding_mat`) REFERENCES `wood_cladding` (`wood_cladding_mat`),
-  CONSTRAINT `fk_request_user1` FOREIGN KEY (`customer_id`) REFERENCES `user` (`user_id`),
-  CONSTRAINT `fk_request_user2` FOREIGN KEY (`admin_id`) REFERENCES `user` (`user_id`)
+                           `request_id` int NOT NULL AUTO_INCREMENT,
+                           `length_cp` int NOT NULL,
+                           `width_cp` int NOT NULL,
+                           `length_rr` int NOT NULL,
+                           `width_rr` int NOT NULL,
+                           `roof_mat` varchar(255) NOT NULL,
+                           `wood_cladding_mat` varchar(255) NOT NULL,
+                           `customer_id` int NOT NULL,
+                           `admin_id` int DEFAULT NULL,
+                           PRIMARY KEY (`request_id`),
+                           KEY `fk_design_roof1_idx` (`roof_mat`),
+                           KEY `fk_design_bekledning1_idx` (`wood_cladding_mat`),
+                           KEY `fk_request_user2_idx` (`admin_id`),
+                           KEY `fk_request_user1_idx` (`customer_id`),
+                           CONSTRAINT `fk_design_roof1` FOREIGN KEY (`roof_mat`) REFERENCES `roof` (`roof_mat`),
+                           CONSTRAINT `fk_design_wood_cladding1` FOREIGN KEY (`wood_cladding_mat`) REFERENCES `wood_cladding` (`wood_cladding_mat`),
+                           CONSTRAINT `fk_request_user1` FOREIGN KEY (`customer_id`) REFERENCES `user` (`user_id`),
+                           CONSTRAINT `fk_request_user2` FOREIGN KEY (`admin_id`) REFERENCES `user` (`user_id`)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb3;
 /*!40101 SET character_set_client = @saved_cs_client */;
 
@@ -89,11 +89,11 @@ DROP TABLE IF EXISTS `roof`;
 /*!40101 SET @saved_cs_client     = @@character_set_client */;
 /*!50503 SET character_set_client = utf8mb4 */;
 CREATE TABLE `roof` (
-  `roof_mat` varchar(255) NOT NULL,
-  `stock_id` int NOT NULL,
-  PRIMARY KEY (`roof_mat`),
-  KEY `fk_roof_stock1_idx` (`stock_id`),
-  CONSTRAINT `fk_roof_stock1` FOREIGN KEY (`stock_id`) REFERENCES `stock` (`stock_id`)
+                        `roof_mat` varchar(255) NOT NULL,
+                        `stock_id` double NOT NULL,
+                        PRIMARY KEY (`roof_mat`),
+                        KEY `fk_roof_stock1_idx` (`stock_id`),
+                        CONSTRAINT `fk_roof_stock1` FOREIGN KEY (`stock_id`) REFERENCES `stock` (`stock_id`)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb3;
 /*!40101 SET character_set_client = @saved_cs_client */;
 
@@ -105,12 +105,12 @@ DROP TABLE IF EXISTS `stock`;
 /*!40101 SET @saved_cs_client     = @@character_set_client */;
 /*!50503 SET character_set_client = utf8mb4 */;
 CREATE TABLE `stock` (
-  `stock_id` int NOT NULL,
-  `description` varchar(255) NOT NULL,
-  `amount` int NOT NULL,
-  `unit` varchar(45) NOT NULL,
-  `price_per_unit` int NOT NULL,
-  PRIMARY KEY (`stock_id`)
+                         `stock_id` double NOT NULL,
+                         `description` varchar(255) NOT NULL,
+                         `amount` int NOT NULL,
+                         `unit` varchar(45) NOT NULL,
+                         `price_per_unit` double NOT NULL,
+                         PRIMARY KEY (`stock_id`)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb3;
 /*!40101 SET character_set_client = @saved_cs_client */;
 
@@ -122,17 +122,17 @@ DROP TABLE IF EXISTS `user`;
 /*!40101 SET @saved_cs_client     = @@character_set_client */;
 /*!50503 SET character_set_client = utf8mb4 */;
 CREATE TABLE `user` (
-  `user_id` int NOT NULL AUTO_INCREMENT,
-  `username` varchar(16) NOT NULL,
-  `email` varchar(255) NOT NULL,
-  `password` varchar(32) NOT NULL,
-  `tlfnr` varchar(20) NOT NULL,
-  `address` varchar(45) DEFAULT NULL,
-  `city` varchar(45) DEFAULT NULL,
-  `isAdmin` tinyint NOT NULL,
-  PRIMARY KEY (`user_id`),
-  KEY `fk_customer_city_idx` (`city`),
-  CONSTRAINT `fk_customer_city` FOREIGN KEY (`city`) REFERENCES `city` (`city`)
+                        `user_id` int NOT NULL AUTO_INCREMENT,
+                        `username` varchar(16) NOT NULL,
+                        `email` varchar(255) NOT NULL,
+                        `password` varchar(32) NOT NULL,
+                        `tlfnr` varchar(20) NOT NULL,
+                        `address` varchar(45) DEFAULT NULL,
+                        `city` varchar(45) DEFAULT NULL,
+                        `isAdmin` tinyint NOT NULL,
+                        PRIMARY KEY (`user_id`),
+                        KEY `fk_customer_city_idx` (`city`),
+                        CONSTRAINT `fk_customer_city` FOREIGN KEY (`city`) REFERENCES `city` (`city`)
 ) ENGINE=InnoDB AUTO_INCREMENT=8 DEFAULT CHARSET=utf8mb3;
 /*!40101 SET character_set_client = @saved_cs_client */;
 
@@ -144,11 +144,11 @@ DROP TABLE IF EXISTS `wood_cladding`;
 /*!40101 SET @saved_cs_client     = @@character_set_client */;
 /*!50503 SET character_set_client = utf8mb4 */;
 CREATE TABLE `wood_cladding` (
-  `wood_cladding_mat` varchar(255) NOT NULL,
-  `stock_id` int NOT NULL,
-  PRIMARY KEY (`wood_cladding_mat`),
-  KEY `fk_bekledning_stock1_idx` (`stock_id`),
-  CONSTRAINT `fk_wood_cladding_stock1` FOREIGN KEY (`stock_id`) REFERENCES `stock` (`stock_id`)
+                                 `wood_cladding_mat` varchar(255) NOT NULL,
+                                 `stock_id` double NOT NULL,
+                                 PRIMARY KEY (`wood_cladding_mat`),
+                                 KEY `fk_bekledning_stock1_idx` (`stock_id`),
+                                 CONSTRAINT `fk_wood_cladding_stock1` FOREIGN KEY (`stock_id`) REFERENCES `stock` (`stock_id`)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb3;
 /*!40101 SET character_set_client = @saved_cs_client */;
 /*!40103 SET TIME_ZONE=@OLD_TIME_ZONE */;
